Adds Go toolchain to streisand host.
This commit adds a `ansible-galaxy` invocation of `galaxy_roles.yml` to
allow depending on external roles. The first such dependency is jlund's
ansible-go role for a go toolchain. Eventually external roles should be
installed in a different directory that can be added singularly to the
.gitignore, for now the role is installed along side the other roles in
the project and we ignore it directly in .gitignore.

Currently jlund/ansible-go is shipping Go 1.7.1 while the latest release
is Go 1.8.1 so pending a PR being merged to update that we override the
defaults in `playbooks/group_vars/all`.