Updates the streisand-client to use the CA, removing --insecure.
This commit updates the client role to copy the generated streisand CA
cert from the provisioning machine to the test machine. The test scripts
are updated so that the --insecure argument being used with curl is
replaced by `--cacert` specifying the streisand CA cert file. This
tests that the gateway TLS validates as expected when the CA has been
installed out-of-band.