Enforce ansible_distribution_version is exactly 16.04
Previous to this commit the `playbooks/roles/common/tasks/main.yml`
block for enforcing that the machine is Ubuntu 16.04 was only checking
the `ansible_distribution` and the `ansible_distribution_major_version`,
meaning that Ubuntu 16.10 was treated the same as Ubuntu 16.04.

This commit reworks the check to enforce that
`ansible_distribution_version` is exactly 16.04.